Updated macro.py
Added try-finally for running the user's functions upon hotkeys being pressed. Previously, a wrapper ran the user's function, and once the function had completed it marked the thread as killable by the outer program. If the function errored, then the wrapper would not reach that code marking the thread as finished.